separated demo entities, updated readme
