+++ README.md
@@ -34,33 +34,37 @@ Pseudo 3d environment generated via ray casting in which the user can explore a 
 - customizable parameters (in `config.json`);
 - custom textures for walls, floor, ceiling and sky;
 - custom billboard sprite rendering in both perspective modes.
+- scriptable entities
 Note: distance based shading of horizontal planes (ceiling/floor) and sky are only available in linear perspective mode.
 
 ## Usage
-Various variables are imported from the json file `config.json` that can be found in the asset folder, usage explained below. A list of game entities (includes both static sprites and moving characters)can be defines as shown in `demo.cpp`.
-The map is generated as a maze using a [randomized DFS](https://en.wikipedia.org/wiki/Maze_generation_algorithm#Randomized_depth-first_search) algorithm, and the process of creation is displayed at game start.
+Various variables are imported from the json file `config.json` that can be found in the asset folder, usage explained below. Scriptable game entities (includes player, static sprites and active entities) can be created and added as shown in `demo.cpp`.
+The maze is generated using a [randomized DFS](https://en.wikipedia.org/wiki/Maze_generation_algorithm#Randomized_depth-first_search) algorithm, and the process of creation is displayed at game start.
 
 ### `assets/config.json` variables
-- playerTransform: initial position and facing direction;
 - gameCamera: field of view in degrees, rays max length, ray precision;
 - gameMap: width, height, option to generate a maze with given dimensions, file form where to load a rectangular map of given (or less) dimensions;
 - screenStats: scale of the minimap (3 means that the center of the m.m. will be at 1/3 of window height, right align), multiplier of the wall height;
 - controls: mouse speed and movement speed;
-- textures: file paths for all textures (alpha);
+- textures: file paths for all textures;
 - sprites: a list of pairs of an id (int) and a path(string) from where the sprite will be loaded.
 
-Note: a sprite size multiplier, an option for using flat colors instead of textures ~~and the possibility of using the same texture for multiple sprites~~ are, as of now, missing features. Ray precision has no effect on the DDA algorithm that is currently in use.
+Note: ~~a sprite size multiplier,~~ an option for using flat colors instead of textures ~~and the possibility of using the same texture for multiple sprites~~ are, as of now, missing features. Ray precision has no effect on the DDA algorithm that is currently in use.
 
 ## Controls
-- WASD to move,
-- mouse left and right or `<` `>` to turn left and right,
-- `ESC` to pause and view map, 
-- `tab` to view full screen map without pausing,
-- `e` to calculate shortest path (will be displayed in full screen map, as of now not implemented for custom maps),
-- `space` to toggle camera plane on and off,
-- `R` to toggle sky on and off (only linear perspective).
+- Built in:
+   - `ESC` to pause and view map, 
+   - `tab` to view full screen map without pausing,
+   - `e` to calculate shortest path (will be displayed in full screen map, as of now not implemented for custom maps),
+   - `space` to toggle camera plane on and off,
+   - `R` to toggle sky on and off (only linear perspective).
+- Available for scripting:
+   - WASD to move,
+   - move mouse left and right or `<` `>` to turn left and right,
+   - mouse left click or `q` is labled as "leftTrigger"
+   - mouse right click or `e` is labled as "rightTrigger"
 
-Note: the game starts paused. Press `ESC` to gain control
+Note: Controls available for scripting are only exposed by the game handler; a possible implementation is shown in the demo project. The game starts paused. Press `ESC` to gain control.
 
 ### 0.1.0
 
@@ -77,16 +81,15 @@ Billboard sprites
 :-------------------------:
 ![sprites](https://github.com/user-attachments/assets/cf18bbb5-7e49-47b8-9718-380f52d8962a)
 
-## Basic class diagram (old)
-![Caster drawio](https://github.com/user-attachments/assets/6165682c-98fd-404e-9333-6a98c0315d25)
+## Basic class diagram
+old | post refactor
+:-------------------------:|:-------------------------:
+![Caster drawio](https://github.com/user-attachments/assets/6165682c-98fd-404e-9333-6a98c0315d25) | to add
 *All not specified relations are 1:1*
 
 ## Known issues
 - mouse speed lowers as the frame rate increases (whilst arrows rotation behaves correctly);
-- drawing sprites has a big impact on frame rate (multi threading not yet implemented in sprite drawing);
-### Structural issues
-- GameGraphics is bloated, this makes reusing code harder;
-- Treating Core and Graphics as separate modules instead of layers would help decoupling.
+- ~~drawing sprites has a big impact on frame rate (multi threading not yet implemented in sprite drawing);~~
 
 ## Resources
  - Textures:
